+++ Maker/src/main/java/com/app/system/TicketManager.java
@@ -3,12 +3,13 @@ package com.app.system;
 import com.app.client.domain.MatchDto;
 import com.app.client.domain.Winner;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 public interface TicketManager {
 
-    double countMultiplierByWinnerType(MatchDto matchDto, Winner winner);
+    BigDecimal countMultiplierByWinnerType(MatchDto matchDto, Winner winner);
 
-    Map<Winner, Double> getPricesForMatch(MatchDto matchDto);
+    Map<Winner, BigDecimal> getPricesForMatch(MatchDto matchDto);
 
 }
